pyproject.toml
[project]
name = "email-classifier-agent"
version = "0.1.0"
description = "AI-powered email classification agent with OpenAI and Gmail API"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.68.0",
    "questionary>=2.0.0",
    "rich>=13.9.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.11.0",
    "pyyaml>=6.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.100.0",
    "msal>=1.28.0",
    "msgraph-sdk>=1.0.0",
    "azure-identity>=1.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
email-classifier = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["agent", "config", "mcp_clients", "models", "ui", "utils"]
include = ["main.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
